requirements/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
Faker==33.3.1
greenlet==3.1.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
mmh3==5.0.1
orjson==3.10.11
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
tb-mqtt-client==1.10.10
typing_extensions==4.12.2
urllib3==2.2.3
